Refactored `file_writer_thread` and `tcp_client_thread` to provide more flexibility in resource sharing

+++ include/io_threads/tls_client.hpp
@@ -30,6 +30,7 @@
 #include "io_threads/tls_client_context.hpp" ///< for io_threads::tls_client_context
 
 #include <cstddef> ///< for size_t
+#include <cstdint> ///< for uint32_t
 #include <memory> ///< for std::shared_ptr
 #include <system_error> ///< for std::error_code
 
@@ -38,15 +39,15 @@ namespace io_threads
 
 /// According to https://datatracker.ietf.org/doc/html/rfc5246#section-6.2.3
 ///   The length MUST NOT exceed 2^14 + 2048
-[[maybe_unused]] constexpr size_t tls1_2_packet_size_limit{16384 + 2048};
+[[maybe_unused]] constexpr uint32_t tls1_2_packet_size_limit{16384 + 2048,};
 
 /// According to https://datatracker.ietf.org/doc/html/rfc8446#section-5.2
 ///   The length MUST NOT exceed 2^14 + 256 bytes
-[[maybe_unused]] constexpr size_t tls1_3_packet_size_limit{16384 + 256};
+[[maybe_unused]] constexpr uint32_t tls1_3_packet_size_limit{16384 + 256,};
 
 /// According to https://datatracker.ietf.org/doc/html/rfc8449#section-1
 ///   Allocating up to 18K of memory for ciphertext is beyond the capacity of some implementations
-[[maybe_unused]] constexpr size_t tls_packet_size_limit{18432};
+[[maybe_unused]] constexpr uint32_t tls_packet_size_limit{18432,};
 
 struct tls_client_session;
